+++ src/components/table/TableHead.vue
@@ -20,8 +20,8 @@ const delegatedProps = computed(() => {
     :class="
       cn(
         props.height || 'h-[38px]',
-        props.padding || 'px-2',
-        'border-b border-stone-300 text-left align-middle font-bold text-muted-foreground [&:has([role=checkbox])]:pr-0',
+        props.padding || 'px-2.5',
+        'border-b border-stone-300 align-middle font-bold text-muted-foreground [&:has([role=checkbox])]:pr-0',
         props.class,
       )
     "